% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LanguageToolR.R
\docType{package}
\name{LanguageToolR-package}
\alias{LanguageToolR}
\alias{LanguageToolR-package}
\title{LanguageToolR: Provides a Wrapper for the LanguageTool CLI Tool for Spelling, Grammar and Language Checking}
\description{
More information about the LanguageTool CLI tool and its options can be found in the LanguageTool Wiki: wiki.languagetool.org.
}
\author{
\strong{Maintainer}: Clemens Schmid \email{clemens@nevrome.de} (\href{https://orcid.org/0000-0003-3448-5715}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Vilmantas Gegzna (\href{https://orcid.org/0000-0002-9500-5167}{ORCID}) [contributor]
}

}
\keyword{internal}
